Tidy up and rename
